Zamiana typu danych int na short, tam gdzie to możliwe

+++ src/main/java/FocusStacking.java
@@ -1,13 +1,13 @@
 public class FocusStacking {
-    private int [][][] image;
-    public FocusStacking(int[][][] focus, int[][][][] input_images){ //to się później zmieni na chara podczas optymalizacji, ale na razie chcę to ze sobą połączyć i int mi jest potrzebny
+    private short [][][] image;
+    public FocusStacking(short[][][] focus, short[][][][] input_images){ //to się później zmieni na chara podczas optymalizacji, ale na razie chcę to ze sobą połączyć i int mi jest potrzebny
         int numberOfPhotos, height, width;
         numberOfPhotos = input_images.length;
         height = input_images[0].length;
         width = input_images[0][0].length;
-        this.image = new int[height][width][4];
-        int [] pixel;
-        int max_focus_value;
+        this.image = new short[height][width][4];
+        short [] pixel;
+        short max_focus_value;
 
         for(int i=0; i<height; i++){
             for(int j=0; j<width; j++){
@@ -23,10 +23,10 @@ public class FocusStacking {
             }
         }
     }
-    public int[] getPixel(int x, int y){
+    public short[] getPixel(int x, int y){
         return this.image[x][y];
     }
-    public int[][][] getImage(){
+    public short[][][] getImage(){
         return this.image;
     }
 }
